Move tests to the same file
