+++ app/src/main/java/com/github/geohunt/app/model/database/api/Location.kt
@@ -96,20 +96,6 @@ data class Location(var latitude: Double = 0.0,
          * Length of the string representing the coarse hash
          */
         const val COARSE_HASH_SIZE = (2 * Long.SIZE_BYTES) / 2
-
-        fun getCoarseHash(location: Location) : String {
-            val crc32 = CRC32()
-
-            // Define a ~11.1km lattice (at the equator)
-            val coarseLatitude = (location.latitude * 10.0).roundToLong()
-            val coarseLongitude = (location.longitude * 10.0).roundToLong()
-
-            val byteBuffer = ByteBuffer.allocate(2 * Long.SIZE_BYTES)
-                    .putLong(0, coarseLatitude)
-                    .putLong(Double.SIZE_BYTES, coarseLongitude)
-            crc32.update(byteBuffer)
-            return crc32.value.toString(36)
-        }
     }
 }
